Post Tags

+++ Tabloid/Controllers/PostController.cs
@@ -1,5 +1,6 @@
 ﻿using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using System.Security.Claims;
 using Tabloid.Data;
 using Tabloid.Models;
 using Tabloid.Repositories;
@@ -12,9 +13,11 @@ namespace Tabloid.Controllers
     public class PostController : ControllerBase
     {
         private readonly PostRepository _postRepository;
+        private readonly UserProfileRepository _userProfileRepository;
         public PostController(ApplicationDbContext context)
         {
             _postRepository = new PostRepository(context);
+            _userProfileRepository = new UserProfileRepository(context);
         }
 
         [HttpGet]
@@ -32,12 +35,25 @@ namespace Tabloid.Controllers
         [HttpGet("{id}")]
         public IActionResult Get(int id)
         {
-            var post = _postRepository.GetApprovedPostById(id);
-            if (post == null)
+            var currentUserProfile = GetCurrentUserProfile();
+            var post = _postRepository.GetById(id);
+            var aprovedPost = _postRepository.GetApprovedPostById(id);
+            if (post == null && aprovedPost == null)
             {
                 return NotFound();
             }
-            return Ok(post);
+            if (aprovedPost != null)
+            {
+                return Ok(aprovedPost);
+            }
+            else if (post != null && post.UserProfileId == currentUserProfile.Id)
+            {
+                return Ok(post);
+            }
+            else
+            {
+                return Unauthorized();
+            }
         }
 
         [HttpPost]
@@ -47,9 +63,48 @@ namespace Tabloid.Controllers
             return CreatedAtAction("Get", new { id = post.Id }, post);
         }
 
+
+        [HttpPost("addtag")]
+        public IActionResult Post(PostTag postTag)
+        {
+            var currentUserProfile = GetCurrentUserProfile();
+            var post = _postRepository.GetById(postTag.PostId);
+
+            if (currentUserProfile.Id != post.UserProfileId)
+            {
+                return Unauthorized();
+            }
+            _postRepository.InsertTag(postTag);
+            return CreatedAtAction("Get", new { id = postTag.Id }, postTag);
+        }
+
+        [HttpDelete("addtag/{id}")]
+        public IActionResult DeletePostTag(int id)
+        {
+            var currentUserProfile = GetCurrentUserProfile();
+            var postTag = _postRepository.GetPostTagById(id);
+
+            var post = _postRepository.GetById(postTag.PostId);
+
+            if (currentUserProfile.Id != post.UserProfileId)
+            {
+                return Unauthorized();
+            }
+
+            _postRepository.RemoveTag(id);
+            return NoContent();
+        }
+
         [HttpPut("{id}")]
         public IActionResult Put(int id, Post post)
         {
+            var currentUserProfile = GetCurrentUserProfile();
+
+            if (currentUserProfile.Id != post.UserProfileId)
+            {
+                return Unauthorized();
+            }
+
             if (id != post.Id)
             {
                 return BadRequest();
@@ -62,8 +117,22 @@ namespace Tabloid.Controllers
         [HttpDelete("{id}")]
         public IActionResult Delete(int id)
         {
+            var currentUserProfile = GetCurrentUserProfile();
+            var post = _postRepository.GetById(id);
+
+            if (currentUserProfile.Id != post.UserProfileId)
+            {
+                return Unauthorized();
+            }
+
             _postRepository.Delete(id);
             return NoContent();
+        }
+
+        private UserProfile GetCurrentUserProfile()
+        {
+            var firebaseUserId = User.FindFirst(ClaimTypes.NameIdentifier).Value;
+            return _userProfileRepository.GetByFirebaseUserId(firebaseUserId);
         }
     }
 }
